back : revu des classes suite maj strucct bdd

+++ Back-end/src/main/java/org/enchere/backend/Repository/CategorieRepository.java
@@ -4,5 +4,5 @@ import org.enchere.backend.model.Categorie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface CategorieRepository extends JpaRepository<Categorie, Long> {
+public interface CategorieRepository extends JpaRepository<Categorie, Integer> {
 }
